Fix front display of attribute settings panel
